bisect--helper: warn if we are assuming an unlikely pathspec
Commit 73c6de06aff8 ("bisect: don't use invalid oid as rev when
starting") changes the behaviour to consider invalid oids as pathspecs
again, as in the old shell implementation.

While that behaviour may be desirable, it can also cause confusion. For
example, while bisecting in a particular repo I encountered this:

    $ git bisect start d93ff48803f0 v6.3
    $

...which led to me sitting for a few moments, wondering why there's no
printout stating the first rev to check.

It turns out that the tag was actually "6.3", not "v6.3", and thus the
bisect was still silently started with only a bad rev, because
d93ff48803f0 was a valid oid and "v6.3" was silently considered to be a
pathspec.

While this behaviour may be desirable, it can be confusing, especially
with different repo conventions either using or not using "v" before
release names, or when a branch name or tag is simply misspelled on the
command line.

In order to avoid this, emit a warning when we are assuming an argument
is a pathspec, but no such path exists in the checkout and this doesn't
look like a pathspec according to looks_like_pathspec:

    $ git bisect start d93ff48803f0 v6.3
    warning: assuming 'v6.3' is a path
    $

If the filename is after the double dash, assume the user knows what
they're doing, just like with other git commands.

Signed-off-by: Chris Down <chris@chrisdown.name>